app/requirements.txt
asgiref==3.8.1
contourpy==1.2.0
cycler==0.12.1
Django==4.1.3
djangorestframework==3.15.1
fonttools==4.50.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
psycopg2

